apps/api/requirements-security.txt
# AGENTLAND.SAARLAND - Security Dependencies
# Production-ready security packages for 200,000 users

# Rate Limiting & DDoS Protection
slowapi==0.1.9
redis==5.0.1

# Input Validation & Sanitization
bleach==6.1.0
python-multipart==0.0.6
email-validator==2.1.0

# Authentication & Authorization
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Security Headers & CORS
fastapi-security==0.5.0

# Cryptography & Encryption
cryptography==41.0.8
pyotp==2.9.0  # For 2FA implementation

# Monitoring & Logging
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
structlog==23.2.0

# Vulnerability Scanning
safety==2.3.5
bandit==1.7.5

# GDPR Compliance
python-gdpr==0.1.0

# IP Geolocation (for security monitoring)
geoip2==4.7.0

# Request/Response Validation
pydantic[email]==2.5.0

# Session Management
itsdangerous==2.1.2

# File Upload Security
python-magic==0.4.27

# Database Security
sqlalchemy-utils==0.41.1

# API Documentation Security
secure-fastapi==1.0.0